backend/requirements.txt
fastapi
uvicorn[standard]
spacy
python-docx
pdfplumber
nltk
chardet
pandas
scikit-learn
joblib